has-a with App abstraction; tests & mypy pass; test clis broken
