+++ src/gwproactor/proactor_implementation.py
@@ -4,15 +4,11 @@ import asyncio
 import sys
 import threading
 import traceback
-import typing
-import uuid
-from dataclasses import dataclass
+from functools import cached_property
 from typing import (
     Any,
-    Callable,
     Dict,
     List,
-    Mapping,
     Optional,
     Sequence,
     Type,
@@ -31,6 +27,12 @@ from paho.mqtt.client import MQTTMessageInfo
 from result import Err, Ok, Result
 
 from gwproactor import ProactorSettings
+from gwproactor.callbacks import (
+    CallbackManager,
+    ProactorCallbackFunctions,
+    ProactorCallbackInterface,
+)
+from gwproactor.config.proactor_config import ProactorConfig, ProactorName
 from gwproactor.external_watchdog import (
     ExternalWatchdogCommandBuilder,
     SystemDWatchdogCommandBuilder,
@@ -40,7 +42,6 @@ from gwproactor.links import (
     AckWaitInfo,
     AsyncioTimerManager,
     LinkManager,
-    Transition,
 )
 from gwproactor.logger import ProactorLogger
 from gwproactor.message import (
@@ -56,7 +57,7 @@ from gwproactor.message import (
     PatWatchdog,
     Shutdown,
 )
-from gwproactor.persister import PersisterInterface, StubPersister
+from gwproactor.persister import PersisterInterface
 from gwproactor.proactor_interface import (
     CommunicatorInterface,
     IOLoopInterface,
@@ -72,31 +73,12 @@ from gwproactor.web_manager import _WebManager
 
 T = TypeVar("T")
 
-PreChildStartCallback = Callable[[], None]
-StartTasksCallback = Callable[[], None]
-StartProcessingMessagesCallback = Callable[[], None]
-ProcessMessageCallback = Callable[[Message[Any]], None]
-ProcessMQTTMessageCallback = Callable[[Message[MQTTReceiptPayload], Message[Any]], None]
-RecvDeactivatedCallback = Callable[[Transition], Result[bool, Exception]]
-RecvActivatedCallback = Callable[[Transition], Result[bool, Exception]]
-
-
-@dataclass
-class ProactorCallbacks:
-    pre_child_start: Optional[PreChildStartCallback] = None
-    start_tasks: Optional[StartTasksCallback] = None
-    start_processing_messages: Optional[StartProcessingMessagesCallback] = None
-    process_message: Optional[ProcessMessageCallback] = None
-    process_mqtt_message: Optional[ProcessMQTTMessageCallback] = None
-    recv_activated: Optional[RecvActivatedCallback] = None
-    recv_deactivated: Optional[RecvDeactivatedCallback] = None
-
 
 class Proactor(ServicesInterface, Runnable):
-    _name: str
+    _name: ProactorName
     _settings: ProactorSettings
     _node: ShNode
-    _callbacks: ProactorCallbacks
+    _callbacks: CallbackManager
     _layout: HardwareLayout
     _logger: ProactorLogger
     _stats: ProactorStats
@@ -113,39 +95,15 @@ class Proactor(ServicesInterface, Runnable):
     _web_manager: _WebManager
     _watchdog: WatchdogManager
 
-    def __init__(
-        self,
-        name: str,
-        settings: ProactorSettings,
-        callbacks: Optional[ProactorCallbacks] = None,
-        hardware_layout: Optional[HardwareLayout] = None,
-    ) -> None:
-        self._name = name
-        self._settings = settings
-        self._callbacks = ProactorCallbacks() if callbacks is None else callbacks
-        if hardware_layout is None:
-            hardware_layout = HardwareLayout(
-                layout={
-                    "ShNodes": [
-                        {
-                            "ShNodeId": str(uuid.uuid4()),
-                            "Name": self._name,
-                            "ActorClass": "NoActor",
-                            "TypeName": "spaceheat.node.gt",
-                        }
-                    ]
-                },
-                cacs={},
-                components={},
-                nodes={},
-                data_channels={},
-                synth_channels={},
-            )
-        self._layout = hardware_layout
-        self._node = self._layout.node(name)
-        self._logger = self.make_logger(settings)
-        self._stats = self.make_stats()
-        self._event_persister = self.make_event_persister(settings)
+    def __init__(self, config: ProactorConfig) -> None:
+        self._name = config.name
+        self._settings = config.settings
+        self._callbacks = CallbackManager(callback_functions=config.callback_functions)
+        self._layout = config.layout
+        self._node = self._layout.node(self.name)
+        self._logger = config.logger
+        self._stats = config.stats
+        self._event_persister = config.event_persister
         self._logger.lifecycle(f"Proactor <{self._name}> reindexing events")
         reindex_result = self._event_persister.reindex()
         self._logger.lifecycle(
@@ -162,7 +120,7 @@ class Proactor(ServicesInterface, Runnable):
         self._links = LinkManager(
             publication_name=self.publication_name,
             subscription_name=self.subscription_name,
-            settings=settings,
+            settings=self._settings,
             logger=self._logger,
             stats=self._stats,
             event_persister=self._event_persister,
@@ -179,27 +137,13 @@ class Proactor(ServicesInterface, Runnable):
         self.add_communicator(self._io_loop_manager)
         self._web_manager = _WebManager(self)
         self.add_communicator(self._web_manager)
-        for config in self._layout.get_components_by_type(WebServerComponent):
+        for server_config in self._layout.get_components_by_type(WebServerComponent):
             self._web_manager.add_web_server_config(
-                name=config.web_server_gt.Name,
-                host=config.web_server_gt.Host,
-                port=config.web_server_gt.Port,
-                **config.web_server_gt.Kwargs,
+                name=server_config.web_server_gt.Name,
+                host=server_config.web_server_gt.Host,
+                port=server_config.web_server_gt.Port,
+                **server_config.web_server_gt.Kwargs,
             )
-
-    @classmethod
-    def make_stats(cls) -> ProactorStats:
-        return ProactorStats()
-
-    @classmethod
-    def make_logger(cls, settings: ProactorSettings) -> ProactorLogger:
-        return ProactorLogger(
-            **typing.cast(Mapping[str, Any], settings.logging.qualified_logger_names())
-        )
-
-    @classmethod
-    def make_event_persister(cls, settings: ProactorSettings) -> PersisterInterface:  # noqa: ARG003
-        return StubPersister()
 
     def send(self, message: Message[Any]) -> None:
         if self._receive_queue is None:
@@ -233,17 +177,33 @@ class Proactor(ServicesInterface, Runnable):
             )
         return communicator
 
-    @property
+    @cached_property
     def name(self) -> str:
-        return self._name
+        return self._name.name
 
-    @property
+    @cached_property
+    def long_name(self) -> str:
+        return self._name.long_name
+
+    @cached_property
+    def short_name(self) -> str:
+        return self._name.short_name
+
+    @cached_property
+    def paths_name(self) -> str:
+        return self._name.paths_name
+
+    @cached_property
     def publication_name(self) -> str:
-        return self._name
+        return self._name.publication_name
+
+    @cached_property
+    def subscription_name(self) -> str:
+        return self._name.subscription_name
 
     @property
-    def subscription_name(self) -> str:
-        return ""
+    def callback_functions(self) -> ProactorCallbackFunctions:
+        return self._callbacks.callback_functions
 
     @property
     def monitored_names(self) -> Sequence[MonitoredName]:
@@ -265,6 +225,12 @@ class Proactor(ServicesInterface, Runnable):
     def links(self) -> LinkManager:
         return self._links
 
+    def add_callbacks(self, callbacks: ProactorCallbackInterface) -> int:
+        return self._callbacks.add_callbacks(callbacks)
+
+    def remove_callbacks(self, callbacks_id: int) -> None:
+        return self._callbacks.remove_callbacks(callbacks_id)
+
     def publish_message(
         self, link_name: str, message: Message[Any], qos: int = 0, context: Any = None
     ) -> MQTTMessageInfo:
@@ -277,6 +243,26 @@ class Proactor(ServicesInterface, Runnable):
     @property
     def io_loop_manager(self) -> IOLoopInterface:
         return self._io_loop_manager
+
+    @property
+    def hardware_layout(self) -> HardwareLayout:
+        return self._layout
+
+    @property
+    def upstream_client(self) -> str:
+        return self._links.upstream_client
+
+    @property
+    def downstream_client(self) -> str:
+        return self._links.downstream_client
+
+    @property
+    def async_receive_queue(self) -> Optional[asyncio.Queue[Any]]:
+        return self._receive_queue
+
+    @property
+    def event_loop(self) -> Optional[asyncio.AbstractEventLoop]:
+        return self._loop
 
     def add_web_server_config(
         self, name: str, host: str, port: int, **kwargs: Any
@@ -303,26 +289,10 @@ class Proactor(ServicesInterface, Runnable):
     def get_web_server_configs(self) -> dict[str, WebServerGt]:
         return self._web_manager.get_configs()
 
-    @property
-    def hardware_layout(self) -> HardwareLayout:
-        return self._layout
-
-    @property
-    def services(self) -> "ServicesInterface":
-        return self
-
     def get_external_watchdog_builder_class(
         self,
     ) -> type[ExternalWatchdogCommandBuilder]:
         return SystemDWatchdogCommandBuilder
-
-    @property
-    def upstream_client(self) -> str:
-        return self._links.upstream_client
-
-    @property
-    def downstream_client(self) -> str:
-        return self._links.downstream_client
 
     def _send(self, message: Message[Any]) -> None:
         self.send(message)
@@ -338,10 +308,7 @@ class Proactor(ServicesInterface, Runnable):
         match self._links.process_ack_timeout(wait_info):
             case Ok(transition):
                 path_dbg |= 0x00000001
-                if (
-                    transition.deactivated()
-                    and self._callbacks.recv_deactivated is not None
-                ):
+                if transition.deactivated():
                     path_dbg |= 0x00000002
                     self._callbacks.recv_deactivated(transition)
             case Err(exception):
@@ -392,22 +359,13 @@ class Proactor(ServicesInterface, Runnable):
         for monitored in communicator.monitored_names:
             self._watchdog.add_monitored_name(monitored)
 
-    @property
-    def async_receive_queue(self) -> Optional[asyncio.Queue[Any]]:
-        return self._receive_queue
-
-    @property
-    def event_loop(self) -> Optional[asyncio.AbstractEventLoop]:
-        return self._loop
-
     async def process_messages(self) -> None:
         if self._receive_queue is None:
             raise RuntimeError(
                 "ERROR. process_messages() called before Proactor started."
             )
         try:
-            if self._callbacks.start_processing_messages is not None:
-                self._callbacks.start_processing_messages()
+            self._callbacks.start_processing_messages()
             while not self._stop_requested:
                 message = await self._receive_queue.get()
                 if not self._stop_requested:
@@ -442,8 +400,7 @@ class Proactor(ServicesInterface, Runnable):
             asyncio.create_task(self.process_messages(), name="process_messages"),
             *self._links.start_ping_tasks(),
         ]
-        if self._callbacks.start_tasks is not None:
-            self._callbacks.start_tasks()
+        self._callbacks.start_tasks()
 
     @classmethod
     def _second_caller(cls) -> str:
@@ -528,9 +485,7 @@ class Proactor(ServicesInterface, Runnable):
                 self.generate_event(message.Payload)
             case _:
                 path_dbg |= 0x00000400
-                if self._callbacks.process_message is not None:
-                    path_dbg |= 0x00000800
-                    self._callbacks.process_message(message)
+                self._callbacks.process_internal_message(message)
         if not isinstance(message.Payload, PatWatchdog):
             self._logger.message_exit(
                 "--Proactor.process_message  path:0x%08X", path_dbg
@@ -600,10 +555,7 @@ class Proactor(ServicesInterface, Runnable):
                 match self._links.process_mqtt_message(mqtt_receipt_message):
                     case Ok(transition):
                         path_dbg |= 0x00000002
-                        if (
-                            transition.recv_activated()
-                            and self._callbacks.recv_activated is not None
-                        ):
+                        if transition.recv_activated():
                             path_dbg |= 0x00000004
                             self._callbacks.recv_activated(transition)
                     case Err(error):
@@ -626,11 +578,9 @@ class Proactor(ServicesInterface, Runnable):
                         self._process_dbg(decoded_message.Payload)
                     case _:
                         path_dbg |= 0x00000080
-                        if self._callbacks.process_mqtt_message is not None:
-                            path_dbg |= 0x00000100
-                            self._callbacks.process_mqtt_message(
-                                mqtt_receipt_message, decoded_message
-                            )
+                        self._callbacks.process_mqtt_message(
+                            mqtt_receipt_message, decoded_message
+                        )
                 if decoded_message.Header.AckRequired:
                     path_dbg |= 0x00000200
                     self._links.send_ack(
@@ -654,11 +604,9 @@ class Proactor(ServicesInterface, Runnable):
         result: Result[bool, Exception] = Ok()
         match self._links.process_mqtt_disconnected(message):
             case Ok(transition):
-                if (
-                    transition.recv_deactivated()
-                    and self._callbacks.recv_deactivated is not None
-                ):
-                    result = self._callbacks.recv_deactivated(transition)
+                if transition.recv_deactivated():
+                    self._callbacks.recv_deactivated(transition)
+                result = Ok(value=True)
             case Err(error):
                 result = Err(error)
         return result
@@ -683,12 +631,10 @@ class Proactor(ServicesInterface, Runnable):
         match self._links.process_mqtt_suback(message):
             case Ok(transition):
                 path_dbg |= 0x00000001
-                if (
-                    transition.recv_activated()
-                    and self._callbacks.recv_activated is not None
-                ):
+                if transition.recv_activated():
                     path_dbg |= 0x00000002
-                    result = self._callbacks.recv_activated(transition)
+                    self._callbacks.recv_activated(transition)
+                    result = Ok(value=True)
             case Err(error):
                 path_dbg |= 0x00000004
                 result = Err(error)
@@ -721,11 +667,6 @@ class Proactor(ServicesInterface, Runnable):
             f"Shutting down due to ShutdownMessage, [{message.Payload.Reason}]"
         )
 
-    def _pre_child_start(self) -> None:
-        """Hook into _start() for derived classes, prior to starting
-        communicators and tasks.
-        """
-
     def _start(self) -> None:
         self._loop = asyncio.get_running_loop()
         self._receive_queue = asyncio.Queue()
@@ -735,7 +676,7 @@ class Proactor(ServicesInterface, Runnable):
                 self._reindex_problems.problem_event("Startup event reindex() problems")
             )
         self._reindex_problems = None
-        self._pre_child_start()
+        self._callbacks.pre_child_start()
         for communicator in self._communicators.values():
             if isinstance(communicator, Runnable):
                 communicator.start()
