+++ src/gwproactor_test/__init__.py
@@ -9,16 +9,8 @@ from gwproactor_test.certs import (
     test_certificate_cache_dir,
 )
 from gwproactor_test.clean import DefaultTestEnv, clean_test_env, default_test_env
-from gwproactor_test.collections import ProactorCommTests
-from gwproactor_test.comm_test_helper import CommTestHelper, ProactorTestHelper
+from gwproactor_test.comm_test_helper import CommTestHelper
 from gwproactor_test.logger_guard import LoggerGuard, LoggerGuards, restore_loggers
-from gwproactor_test.recorder import (
-    ProactorT,
-    RecorderInterface,
-    RecorderLinkStats,
-    RecorderStats,
-    make_recorder_class,
-)
 from gwproactor_test.wait import (
     AwaitablePredicate,
     ErrorStringFunction,
@@ -38,18 +30,11 @@ __all__ = [
     "LoggerGuard",
     "LoggerGuards",
     "Predicate",
-    "ProactorCommTests",
-    "ProactorT",
-    "ProactorTestHelper",
-    "RecorderInterface",
-    "RecorderLinkStats",
-    "RecorderStats",
     "StopWatch",
     "await_for",
     "clean_test_env",
     "copy_keys",
     "default_test_env",
-    "make_recorder_class",
     "restore_loggers",
     "set_test_certificate_cache_dir",
     "test_ca_certificate_path",
