+++ src/gwproactor_test/dummies/__init__.py
@@ -11,9 +11,9 @@ from gwproactor_test.dummies.names import (
     DUMMY_SCADA2_ENV_PREFIX,
     DUMMY_SCADA2_NAME,
 )
-from gwproactor_test.dummies.pair.child import DummyChild
+from gwproactor_test.dummies.pair.child import DummyChildApp
 from gwproactor_test.dummies.pair.child_config import DummyChildSettings
-from gwproactor_test.dummies.pair.parent import DummyParent
+from gwproactor_test.dummies.pair.parent import ParentApp
 from gwproactor_test.dummies.pair.parent_config import DummyParentSettings
 
 __all__ = [
@@ -22,9 +22,9 @@ __all__ = [
     "DUMMY_ENV_PREFIX",
     "DUMMY_PARENT_ENV_PREFIX",
     "DUMMY_PARENT_NAME",
-    "DummyChild",
+    "DummyChildApp",
     "DummyChildSettings",
-    "DummyParent",
+    "ParentApp",
     "DummyParentSettings",
     "DUMMY_SCADA1_ENV_PREFIX",
     "DUMMY_SCADA2_ENV_PREFIX",
