+++ src/gwproactor_test/dummies/pair/__init__.py
@@ -1,11 +1,11 @@
-from gwproactor_test.dummies.pair.child import DummyChild
+from gwproactor_test.dummies.pair.child import DummyChildApp
 from gwproactor_test.dummies.pair.child_config import DummyChildSettings
-from gwproactor_test.dummies.pair.parent import DummyParent
+from gwproactor_test.dummies.pair.parent import ParentApp
 from gwproactor_test.dummies.pair.parent_config import DummyParentSettings
 
 __all__ = [
-    "DummyChild",
+    "DummyChildApp",
     "DummyChildSettings",
-    "DummyParent",
+    "ParentApp",
     "DummyParentSettings",
 ]
