+++ src/gwproactor_test/dummies/pair/child.py
@@ -1,20 +1,18 @@
-import typing
-from typing import Optional
-
 from gwproto import Message, MQTTCodec, MQTTTopic, create_message_model
 
-from gwproactor import ProactorSettings
+from gwproactor import Proactor, ProactorSettings
+from gwproactor.config import MQTTClient
+from gwproactor.config.proactor_config import ProactorName
 from gwproactor.links import QOS
 from gwproactor.links.link_settings import LinkSettings
-from gwproactor.persister import TimedRollingFilePersister
-from gwproactor.proactor_implementation import Proactor, ProactorCallbacks
+from gwproactor.persister import PersisterInterface, TimedRollingFilePersister
 from gwproactor_test.dummies.names import (
     CHILD_SHORT_NAME,
     DUMMY_CHILD_NAME,
     DUMMY_PARENT_NAME,
     PARENT_SHORT_NAME,
 )
-from gwproactor_test.dummies.pair.child_config import DummyChildSettings
+from gwproactor_test.instrumented_app import TestApp
 
 
 class ChildMQTTCodec(MQTTCodec):
@@ -38,26 +36,32 @@ class ChildMQTTCodec(MQTTCodec):
             )
 
 
-class DummyChild(Proactor):
-    PARENT_MQTT = "gridworks"
+class DummyChildApp(TestApp):
+    PARENT_MQTT: str = "parent"
 
-    def __init__(
-        self,
-        name: str = "",
-        settings: Optional[DummyChildSettings] = None,
-        callbacks: Optional[ProactorCallbacks] = None,
-    ) -> None:
-        super().__init__(
-            name=name or DUMMY_CHILD_NAME,
-            settings=DummyChildSettings() if settings is None else settings,
-            callbacks=callbacks,
+    @classmethod
+    def get_name(cls) -> ProactorName:
+        return ProactorName(
+            long_name=DUMMY_CHILD_NAME,
+            short_name=CHILD_SHORT_NAME,
+            paths_name=DUMMY_CHILD_NAME,
         )
-        self._links.add_mqtt_link(
+
+    @classmethod
+    def get_mqtt_clients(cls) -> dict[str, MQTTClient]:
+        return {DummyChildApp.PARENT_MQTT: MQTTClient()}
+
+    @classmethod
+    def make_persister(cls, settings: ProactorSettings) -> PersisterInterface:
+        return TimedRollingFilePersister(settings.paths.event_dir)
+
+    def connect_proactor(self, proactor: Proactor) -> None:
+        proactor.links.add_mqtt_link(
             LinkSettings(
-                client_name=DummyChild.PARENT_MQTT,
+                client_name=DummyChildApp.PARENT_MQTT,
                 gnode_name=DUMMY_PARENT_NAME,
                 spaceheat_name=PARENT_SHORT_NAME,
-                mqtt=self.settings.parent_mqtt,
+                mqtt=proactor.settings.mqtt[self.PARENT_MQTT],
                 codec=ChildMQTTCodec(),
                 upstream=True,
             )
@@ -66,23 +70,4 @@ class DummyChild(Proactor):
             MQTTTopic.encode_subscription(Message.type_name(), "1", "a"),
             MQTTTopic.encode_subscription(Message.type_name(), "2", "b"),
         ]:
-            self._links.subscribe(self.PARENT_MQTT, topic, QOS.AtMostOnce)
-        self._links.log_subscriptions("construction")
-
-    @classmethod
-    def make_event_persister(
-        cls, settings: ProactorSettings
-    ) -> TimedRollingFilePersister:
-        return TimedRollingFilePersister(settings.paths.event_dir)
-
-    @property
-    def publication_name(self) -> str:
-        return self.name
-
-    @property
-    def subscription_name(self) -> str:
-        return CHILD_SHORT_NAME
-
-    @property
-    def settings(self) -> DummyChildSettings:
-        return typing.cast(DummyChildSettings, self._settings)
+            proactor.links.subscribe(self.PARENT_MQTT, topic, QOS.AtMostOnce)
