+++ src/gwproactor_test/dummies/pair/parent.py
@@ -1,19 +1,24 @@
 """Scada implementation"""
 
-from typing import Optional, cast
+from typing import Any, Optional
 
-from gwproto import MQTTCodec, create_message_model
+from gwproto import Message, MQTTCodec, create_message_model
+from result import Ok, Result
 
+from gwproactor import Proactor, ProactorSettings
+from gwproactor.actors.actor import PrimeActor
+from gwproactor.config import MQTTClient
+from gwproactor.config.proactor_config import ProactorConfig, ProactorName
 from gwproactor.links.link_settings import LinkSettings
-from gwproactor.persister import SimpleDirectoryWriter
-from gwproactor.proactor_implementation import Proactor, ProactorCallbacks
+from gwproactor.message import DBGPayload
+from gwproactor.persister import PersisterInterface, SimpleDirectoryWriter
 from gwproactor_test.dummies.names import (
     CHILD_SHORT_NAME,
     DUMMY_CHILD_NAME,
     DUMMY_PARENT_NAME,
     PARENT_SHORT_NAME,
 )
-from gwproactor_test.dummies.pair.parent_config import DummyParentSettings
+from gwproactor_test.instrumented_app import TestApp
 
 
 class ParentMQTTCodec(MQTTCodec):
@@ -34,46 +39,60 @@ class ParentMQTTCodec(MQTTCodec):
             )
 
 
-class DummyParent(Proactor):
-    CHILD_MQTT = "child"
+class DummyParent(PrimeActor):
+    def process_internal_message(self, message: Message[Any]) -> None:
+        self.process_message(message)
+
+    def process_message(self, message: Message[Any]) -> Result[bool, Exception]:
+        match message.Payload:
+            case DBGPayload():
+                message.Header.Src = self.services.publication_name
+                dst_client = message.Header.Dst
+                message.Header.Dst = ""
+                self.services.publish_message(dst_client, message)
+        return Ok(True)
+
+
+class ParentApp(TestApp):
+    CHILD_MQTT: str = "child"
 
     def __init__(
         self,
-        name: str = "",
-        settings: Optional[DummyParentSettings] = None,
-        callbacks: Optional[ProactorCallbacks] = None,
+        *,
+        name: Optional[ProactorName] = None,
+        settings: Optional[ProactorSettings] = None,
+        config: Optional[ProactorConfig] = None,
     ) -> None:
         super().__init__(
-            name=name or DUMMY_PARENT_NAME,
-            settings=DummyParentSettings() if settings is None else settings,
-            callbacks=callbacks,
+            name=name, settings=settings, config=config, prime_actor_type=DummyParent
         )
-        self._links.add_mqtt_link(
+
+    @classmethod
+    def get_name(cls) -> ProactorName:
+        return ProactorName(
+            long_name=DUMMY_PARENT_NAME,
+            short_name=PARENT_SHORT_NAME,
+            paths_name=DUMMY_PARENT_NAME,
+        )
+
+    @classmethod
+    def get_mqtt_clients(cls) -> dict[str, MQTTClient]:
+        return {ParentApp.CHILD_MQTT: MQTTClient()}
+
+    @classmethod
+    def make_persister(cls, settings: ProactorSettings) -> PersisterInterface:
+        return SimpleDirectoryWriter(settings.paths.event_dir)
+
+    def connect_proactor(self, proactor: Proactor) -> None:
+        if self.proactor is None:
+            raise RuntimeError("connect_proactor called before instantiate()")
+        proactor.links.add_mqtt_link(
             LinkSettings(
                 client_name=self.CHILD_MQTT,
                 gnode_name=DUMMY_CHILD_NAME,
                 spaceheat_name=CHILD_SHORT_NAME,
-                mqtt=self.settings.child_mqtt,
+                mqtt=proactor.settings.mqtt[self.CHILD_MQTT],
                 codec=ParentMQTTCodec(),
                 downstream=True,
             )
         )
-
-    @classmethod
-    def make_event_persister(
-        cls,
-        settings: DummyParentSettings,  # type: ignore[override]
-    ) -> SimpleDirectoryWriter:
-        return SimpleDirectoryWriter(settings.paths.event_dir)
-
-    @property
-    def publication_name(self) -> str:
-        return self.name
-
-    @property
-    def settings(self) -> DummyParentSettings:
-        return cast(DummyParentSettings, self._settings)
-
-    @property
-    def subscription_name(self) -> str:
-        return PARENT_SHORT_NAME
