+++ src/gwproactor_test/dummies/tree/atn1_cli.py
@@ -7,7 +7,7 @@ from gwproactor.command_line_utils import (
     run_async_main,
 )
 from gwproactor_test.dummies import DUMMY_ATN_NAME
-from gwproactor_test.dummies.tree.atn import DummyAtn
+from gwproactor_test.dummies.tree.atn import DummyAtnApp
 from gwproactor_test.dummies.tree.atn_settings import DummyAtnSettings
 
 app = typer.Typer(
@@ -25,10 +25,11 @@ def run(
     verbose: bool = False,
     message_summary: bool = False,
 ) -> None:
-    asyncio.run(
+    raise NotImplementedError("Reimplement CLIs using App class")
+    asyncio.run(  # type: ignore[unreachable]
         run_async_main(
             name=DUMMY_ATN_NAME,
-            proactor_type=DummyAtn,
+            proactor_type=DummyAtnApp,
             settings_type=DummyAtnSettings,
             env_file=env_file,
             dry_run=dry_run,
