+++ src/gwproactor_test/dummies/tree/scada2.py
@@ -5,86 +5,48 @@ from typing import Optional
 import rich
 from gwproto import Message
 
-from gwproactor import Proactor, ProactorSettings
+from gwproactor import Proactor, ProactorSettings, ServicesInterface
+from gwproactor.actors.actor import PrimeActor
+from gwproactor.config import MQTTClient
+from gwproactor.config.proactor_config import ProactorConfig, ProactorName
 from gwproactor.links.link_settings import LinkSettings
 from gwproactor.message import MQTTReceiptPayload
 from gwproactor.persister import TimedRollingFilePersister
-from gwproactor.proactor_implementation import ProactorCallbacks
+from gwproactor_test.dummies import DUMMY_SCADA2_NAME
 from gwproactor_test.dummies.names import DUMMY_SCADA2_SHORT_NAME
 from gwproactor_test.dummies.tree.admin_messages import (
     AdminCommandSetRelay,
     AdminSetRelayEvent,
 )
+from gwproactor_test.dummies.tree.admin_settings import AdminLinkSettings
 from gwproactor_test.dummies.tree.codecs import AdminCodec, DummyCodec
 from gwproactor_test.dummies.tree.messages import (
     RelayInfo,
     RelayReportEvent,
     SetRelay,
 )
-from gwproactor_test.dummies.tree.scada2_settings import DummyScada2Settings
+from gwproactor_test.dummies.tree.scada1_settings import AtnLinkSettings
+from gwproactor_test.dummies.tree.scada2_settings import (
+    Scada1LinkSettings,
+)
+from gwproactor_test.instrumented_app import TestApp
 
 
-class DummyScada2(Proactor):
+class DummyScada2(PrimeActor):
+    ADMIN_LINK: str = "admin_link"
+
     relays: dict[str, bool]
 
-    def __init__(
-        self,
-        name: str = "",
-        settings: Optional[DummyScada2Settings] = None,
-        callbacks: Optional[ProactorCallbacks] = None,
-    ) -> None:
-        if settings is None:
-            settings = DummyScada2Settings()
-        self.relays = defaultdict(bool)
-        callbacks = ProactorCallbacks() if callbacks is None else callbacks
-        callbacks.process_mqtt_message = self._derived_process_mqtt_message
-        super().__init__(name=name, settings=settings, callbacks=callbacks)
-        self._links.add_mqtt_link(
-            LinkSettings(
-                client_name=self.settings.scada1_link.client_name,
-                gnode_name=self.settings.scada1_link.long_name,
-                spaceheat_name=self.settings.scada1_link.short_name,
-                mqtt=self.settings.scada1_link,
-                codec=DummyCodec(
-                    src_name=self.settings.scada1_link.long_name,
-                    dst_name=DUMMY_SCADA2_SHORT_NAME,
-                    model_name="Scada1ToScada2Message",
-                ),
-                upstream=True,
-            ),
-        )
-        if self.settings.admin_link.enabled:
-            self._links.add_mqtt_link(
-                LinkSettings(
-                    client_name=self.settings.admin_link.client_name,
-                    gnode_name=self.settings.admin_link.long_name,
-                    spaceheat_name=self.settings.admin_link.short_name,
-                    mqtt=self.settings.admin_link,
-                    codec=AdminCodec(),
-                ),
-            )
-        self.links.log_subscriptions("construction")
-
-    @property
-    def settings(self) -> DummyScada2Settings:
-        return typing.cast(DummyScada2Settings, self._settings)
-
-    @property
-    def subscription_name(self) -> str:
-        return DUMMY_SCADA2_SHORT_NAME
+    def __init__(self, name: str, services: ServicesInterface) -> None:
+        super().__init__(name, services)
+        self.relays = self.relays = defaultdict(bool)
 
     @property
     def admin_client(self) -> str:
-        return self.settings.admin_link.client_name
-
-    @classmethod
-    def make_event_persister(
-        cls, settings: ProactorSettings
-    ) -> TimedRollingFilePersister:
-        return TimedRollingFilePersister(settings.paths.event_dir)
+        return self.ADMIN_LINK
 
     def _process_set_relay(self, payload: RelayInfo) -> None:
-        self._logger.path(
+        self.services.logger.path(
             f"++{self.name}._process_set_relay "
             f"{payload.RelayName}  "
             f"closed:{payload.Closed}"
@@ -99,8 +61,8 @@ class DummyScada2(Proactor):
         if event.changed:
             path_dbg |= 0x00000001
             self.relays[payload.RelayName] = event.closed
-        self.generate_event(event)
-        self._logger.path(
+        self.services.generate_event(event)
+        self.services.logger.path(
             f"--{self.name}._process_set_relay  "
             f"path:0x{path_dbg:08X}  "
             f"{int(last_val)} -> "
@@ -110,7 +72,7 @@ class DummyScada2(Proactor):
     def _process_upstream_mqtt_message(
         self, message: Message[MQTTReceiptPayload], decoded: Message[typing.Any]
     ) -> None:
-        self._logger.path(
+        self.services.logger.path(
             f"++{self.name}._process_downstream_mqtt_message {message.Payload.message.topic}",
         )
         path_dbg = 0
@@ -125,21 +87,21 @@ class DummyScada2(Proactor):
                     f"There is no handler for mqtt message payload type [{type(decoded.Payload)}]\n"
                     f"Received\n\t topic: [{message.Payload.message.topic}]"
                 )
-        self._logger.path(
+        self.services.logger.path(
             f"--{self.name}._process_downstream_mqtt_message  path:0x{path_dbg:08X}",
         )
 
     def _process_admin_mqtt_message(
         self, message: Message[MQTTReceiptPayload], decoded: Message[typing.Any]
     ) -> None:
-        self._logger.path(
+        self.services.logger.path(
             f"++{self.name}._process_admin_mqtt_message {message.Payload.message.topic}",
         )
         path_dbg = 0
         match decoded.Payload:
             case AdminCommandSetRelay() as command:
                 path_dbg |= 0x00000001
-                self.generate_event(AdminSetRelayEvent(command=command))
+                self.services.generate_event(AdminSetRelayEvent(command=command))
                 self._process_set_relay(command.RelayInfo)
             case _:
                 raise ValueError(
@@ -149,18 +111,18 @@ class DummyScada2(Proactor):
                     f"Received\n\t topic: [{message.Payload.message.topic}]"
                 )
 
-        self._logger.path(
+        self.services.logger.path(
             f"--{self.name}._process_admin_mqtt_message  path:0x{path_dbg:08X}",
         )
 
-    def _derived_process_mqtt_message(
+    def process_mqtt_message(
         self, message: Message[MQTTReceiptPayload], decoded: Message[typing.Any]
     ) -> None:
-        self._logger.path(
+        self.services.logger.path(
             f"++{self.name}._derived_process_mqtt_message {message.Payload.message.topic}",
         )
         path_dbg = 0
-        if message.Payload.client_name == self.upstream_client:
+        if message.Payload.client_name == self.services.upstream_client:
             path_dbg |= 0x00000001
             self._process_upstream_mqtt_message(message, decoded)
         elif message.Payload.client_name == self.admin_client:
@@ -174,6 +136,82 @@ class DummyScada2(Proactor):
                 f"[{message.Payload.client_name}]\n"
                 f"Received\n\t topic: [{message.Payload.message.topic}]"
             )
-        self._logger.path(
+        self.services.logger.path(
             f"--{self.name}._derived_process_mqtt_message  path:0x{path_dbg:08X}",
         )
+
+
+class DummyScada2App(TestApp):
+    ATN_LINK: str = "atn_link"
+    SCADA1_LINK: str = "scada1_link"
+
+    atn_link: AtnLinkSettings
+    scada1_link: Scada1LinkSettings
+    admin_link: AdminLinkSettings
+
+    def __init__(
+        self,
+        *,
+        name: Optional[ProactorName] = None,
+        settings: Optional[ProactorSettings] = None,
+        config: Optional[ProactorConfig] = None,
+    ) -> None:
+        super().__init__(
+            name=name, settings=settings, config=config, prime_actor_type=DummyScada2
+        )
+        self.atn_link = AtnLinkSettings(
+            **self.config.settings.mqtt[self.ATN_LINK].model_dump()
+        )
+        self.scada1_link = Scada1LinkSettings(
+            **self.config.settings.mqtt[self.SCADA1_LINK].model_dump()
+        )
+        self.admin_link = AdminLinkSettings(
+            **self.config.settings.mqtt[DummyScada2.ADMIN_LINK].model_dump()
+        )
+
+    @classmethod
+    def get_name(cls) -> ProactorName:
+        return ProactorName(
+            long_name=DUMMY_SCADA2_NAME,
+            short_name=DUMMY_SCADA2_SHORT_NAME,
+            paths_name=DUMMY_SCADA2_NAME,
+        )
+
+    @classmethod
+    def get_mqtt_clients(cls) -> dict[str, MQTTClient]:
+        return {
+            cls.ATN_LINK: MQTTClient(),
+            cls.SCADA1_LINK: MQTTClient(),
+            DummyScada2.ADMIN_LINK: MQTTClient(),
+        }
+
+    @classmethod
+    def make_persister(cls, settings: ProactorSettings) -> TimedRollingFilePersister:
+        return TimedRollingFilePersister(settings.paths.event_dir)
+
+    def connect_proactor(self, proactor: Proactor) -> None:
+        proactor.links.add_mqtt_link(
+            LinkSettings(
+                client_name=self.scada1_link.client_name,
+                gnode_name=self.scada1_link.long_name,
+                spaceheat_name=self.scada1_link.short_name,
+                mqtt=self.scada1_link,
+                codec=DummyCodec(
+                    src_name=self.scada1_link.long_name,
+                    dst_name=DUMMY_SCADA2_SHORT_NAME,
+                    model_name="Scada1ToScada2Message",
+                ),
+                upstream=True,
+            ),
+        )
+        if self.admin_link.enabled:
+            proactor.links.add_mqtt_link(
+                LinkSettings(
+                    client_name=self.admin_link.client_name,
+                    gnode_name=self.admin_link.long_name,
+                    spaceheat_name=self.admin_link.short_name,
+                    mqtt=self.admin_link,
+                    codec=AdminCodec(),
+                ),
+            )
+        proactor.links.log_subscriptions("construction")
