+++ src/gwproactor_test/logger_guard.py
@@ -1,3 +1,5 @@
+# ruff: noqa: S101
+
 import logging
 import sys
 from types import TracebackType
